Project BASHO/basho-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.25.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dateutil==2.8.2
httpx==0.25.1feedparser==6.0.10